gcp_cost_control/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.30.0
cloudevents==1.*
